requirements.txt
backports.functools-lru-cache==1.6.4
bayesian-optimization==1.2.0
blinker==1.4
boto3==1.24.57
brotlipy==0.7.0
cached-property==1.5.2
colorama==0.4.5
fastapi==0.80.0
flask==2.2.2
flit-core==3.7.1
fonttools==4.37.0
gym==0.25.2
importlib-resources==5.9.0
jupyter==1.0.0
kaggle==1.5.12
keras==2.9.0
keras-balanced-batch-generator==0.0.2
matplotlib==3.1.2
munkres==1.1.4
nb-conda==2.2.1
nb-conda-kernels==2.3.1
pandas-datareader==0.10.0
pillow==9.2.0
pip-chill==1.0.1
pkgutil-resolve-name==1.3.10
pyopenssl==22.0.0
pysocks==1.7.1
python-multipart==0.0.5
pyyaml==6.0
rich==12.5.1
tensorflow-addons==0.17.1
tensorflow-macos==2.5.0
tensorflow-metal==0.1.2
unicodedata2==14.0.0
uvicorn==0.18.2
